🔨 rename GdocsConstants -> GdocConstants
